Update project dependencies and GitHub workflows
This major update includes an upgrade of the sbt version to 1.10.0 and several plugin versions in the project. The Java version used in GitHub actions has been changed from 8 to 17. Additionally, the Scala version was updated from 2.13.11 to 2.13.14 in the build.sbt file. Furthermore, the GitHub workflows have been updated to use newer versions of actions and a Dependabot configuration file has been added to the project.